refactor: move get_changed_items logic to _common/_git_utils.py
